Fix debugger bug
If the debugger menu branches off to logic that changes the command variable
from a string to a list, then subsequent conditional checks throw an error
when they try to access the "startswith" method on that command variable.
Changing the "ifs" to "elifs" like they should have been originally will
prevent this from happening.